requirement.txt
Pillow
scikit-learn
torch
torchvision
opencv-python
pyzmq
